implemented search feature
Removed the textfield and added a button that allows users to search for their ward. It would be optimal if users could search using that textfield and for the cards to slide up to the top but due to time constraints I stuck to this method for now. At the end, if there is time for me or Jimin then we can upgrade this part.