started creating abouts page
Not nearly close to being done. This about page will contain steps on how to use the app and maybe a little bit of information about us developers as well. I also need to change the UI for the camera page as well but im not sure what actually comes below the camera and photo library icon.